@@ -1,23 +1,26 @@
-import { useRouter } from 'next/router'
+'use client'
+// this is a client component because it uses the `useState` hook
+
 import { useState } from 'react'
-import { t, msg } from '@lingui/macro'
-import { MessageDescriptor } from '@lingui/core/src'
+import { msg } from '@lingui/macro'
 import { useLingui } from '@lingui/react'
+import { usePathname, useRouter } from 'next/navigation'
 
 type LOCALES = 'en' | 'sr' | 'es' | 'pseudo'
 
-const languages: { [key: string]: MessageDescriptor } = {
+const languages = {
   en: msg`English`,
   sr: msg`Serbian`,
   es: msg`Spanish`
-}
+} as const
 
 export function Switcher() {
   const router = useRouter()
   const { i18n } = useLingui()
+  const pathname = usePathname()
 
   const [locale, setLocale] = useState<LOCALES>(
-    router.locale!.split('-')[0] as LOCALES
+    pathname?.split('/')[1] as LOCALES
   )
 
   // disabled for DEMO - so we can demonstrate the 'pseudo' locale functionality
@@ -28,8 +31,11 @@ export function Switcher() {
   function handleChange(event: React.ChangeEvent<HTMLSelectElement>) {
     const locale = event.target.value as LOCALES
 
+    const pathNameWithoutLocale = pathname?.split('/')?.slice(2) ?? []
+    const newPath = `/${locale}/${pathNameWithoutLocale.join('/')}`
+
     setLocale(locale)
-    router.push(router.pathname, router.pathname, { locale })
+    router.push(newPath)
   }
 
   return (
@@ -37,7 +43,7 @@ export function Switcher() {
       {Object.keys(languages).map((locale) => {
         return (
           <option value={locale} key={locale}>
-            {i18n._(languages[locale as unknown as LOCALES])}
+            {i18n._(languages[locale as keyof typeof languages])}
           </option>
         )
       })}
