@@ -1,10 +1,8 @@
-const nextConfig = require('./next.config')
-
 /** @type {import('@lingui/conf').LinguiConfig} */
 module.exports = {
-  locales: nextConfig.i18n.locales,
+  locales: ['en', 'sr', 'es', 'pseudo'],
   pseudoLocale: 'pseudo',
-  sourceLocale: nextConfig.i18n.defaultLocale,
+  sourceLocale: 'en',
   fallbackLocales: {
     default: 'en'
   },
