@@ -1,6 +1,6 @@
 import { i18n, Messages } from '@lingui/core'
-import { useRouter } from 'next/router'
 import { useEffect } from 'react'
+import { usePathname } from 'next/navigation'
 
 export async function loadCatalog(locale: string) {
   const catalog = await import(`@lingui/loader!./locales/${locale}.po`)
@@ -8,9 +8,9 @@ export async function loadCatalog(locale: string) {
 }
 
 export function useLinguiInit(messages: Messages) {
-  const router = useRouter()
-  const locale = router.locale || router.defaultLocale!
   const isClient = typeof window !== 'undefined'
+  const pathname = usePathname()
+  const locale = pathname?.split('/')[1] ?? 'en'
 
   if (!isClient && locale !== i18n.locale) {
     // there is single instance of i18n on the server
